kv: Move raftstore impls in
